fix(related): relocate error function
